+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2021-02-03
+### Added
+- Add filter: camelize
+
+### Changed
+- Change filters: TrimLeft & TrimRight (add specified char to trim)
+- Update README.md
+
 ## [0.2.0] - 2021-02-03
 ### Added
 - Convert & return value as string, int, float & boolean